# Invalidate column name caches when the schema of table may change